+++ examples/dailyhorseracing.py
@@ -83,70 +83,88 @@ except IndexError:
 
 # Get active selections and prices for active fixture
 for fixture in active_fixtures[-1:]:
-    fixture_selection_prices = client.betting.selections_for_market(
-        fixture_id=fixture.fixture_id,
-        market_type_id=win_market_type.market_type_id,
-        top_price_only=False,
-    )
-    top_price_fixture_selection_prices = client.betting.selections_for_market(
-        fixture_id=fixture.fixture_id,
-        market_type_id=win_market_type.market_type_id,
-        top_price_only=True,
-    )
-    if fixture_selection_prices:
-        # select the first selection/competitor with a price and that is active
-        try:
-            selection = [
-                c
-                for c in fixture_selection_prices
-                if (c.trading_status == "Trading") and c.prices
-            ][0]
-        except IndexError as e:
-            raise Exception(
-                "No competitors in the fixture are active and have prices available"
-            )
 
-        try:
-            best_price = [p for p in selection.prices][0]
-        except IndexError as e:
-            raise Exception(f"No price was available for selection {selection.name}")
+    # get active markets
+    active_markets = client.betting.active_markets(fixture_id=fixture.fixture_id)
 
-        # Create a BACK order against the first runner in the market
-        bet_create_response = client.betting.bet_request_create(
-            request_filter=resources.filters.CreateBetRequestFilter(
-                fixture_id=fixture.fixture_id,
-                market_type_id=win_market_type.market_type_id,
-                competitor=selection.competitor_id,
-                price=best_price.price,
-                stake=500,
-                bet_type="Win",
-                customer_strategy_ref=resources.CustomerStrategyRef.create_customer_strategy_ref(
-                    STRATEGY_NAME
-                ),
-                customer_order_ref=resources.CustomerOrderRef.create_customer_order_ref(
-                    str(uuid.uuid4())
-                ),
-            )
+    if active_markets:
+        try:
+            win_bet_type = active_markets[0].bet_types[0]
+        except IndexError:
+            raise Exception(f"Cannot find any win types for the active market")
+
+        # get the selections and prices for a given market
+        fixture_selection_prices = client.betting.selections_for_market(
+            fixture_id=fixture.fixture_id,
+            market_type_id=win_market_type.market_type_id,
+            top_price_only=False,
         )
-        if isinstance(bet_create_response, resources.BetRequestCreate):
-            logger.info(f"Bet request was created!")
 
-            bet_request_get = client.betting.bet_request_get(
-                request_filter=resources.GetBetRequestFilter(
-                    bet_request_id=bet_create_response.bet_request_id
+        # get the top priced selections and prices for a given market
+        top_price_fixture_selection_prices = client.betting.selections_for_market(
+            fixture_id=fixture.fixture_id,
+            market_type_id=win_market_type.market_type_id,
+            top_price_only=True,
+        )
+        if fixture_selection_prices:
+            # select the first selection/competitor with a price and that is active
+            try:
+                selection = [
+                    c
+                    for c in fixture_selection_prices
+                    if (c.trading_status == "Trading") and c.prices
+                ][0]
+            except IndexError as e:
+                raise Exception(
+                    "No competitors in the fixture are active and have prices available"
+                )
+
+            try:
+                best_price = [p for p in selection.prices][0]
+            except IndexError as e:
+                raise Exception(
+                    f"No price was available for selection {selection.name}"
+                )
+
+            # Create a BACK order against the first runner in the market
+            bet_create_response = client.betting.bet_request_create(
+                request_filter=resources.filters.CreateBetRequestFilter(
+                    fixture_id=fixture.fixture_id,
+                    market_type_id=win_market_type.market_type_id,
+                    competitor=selection.competitor_id,
+                    price=best_price.price,
+                    stake=500,
+                    bet_type=win_bet_type,
+                    customer_strategy_ref=resources.CustomerStrategyRef.create_customer_strategy_ref(
+                        STRATEGY_NAME
+                    ),
+                    customer_order_ref=resources.CustomerOrderRef.create_customer_order_ref(
+                        str(uuid.uuid4())
+                    ),
                 )
             )
+            if isinstance(bet_create_response, resources.BetRequestCreate):
+                logger.info(f"Bet request was created!")
 
-            # Retrieve my bets
-            active_bet_request = client.betting.get_active_bet_requests()
+                # get the previously created bet request
+                bet_request_get = client.betting.bet_request_get(
+                    request_filter=resources.GetBetRequestFilter(
+                        bet_request_id=bet_create_response.bet_request_id
+                    )
+                )
 
-            # Get active bets
-            my_active_back_bets = client.betting.my_bets(
-                side=enums.BetSide.BACK, status=enums.BetRequestStatus.ACTIVE
-            )
+                # Retrieve my bets
+                active_bet_request = client.betting.get_active_bet_requests()
+
+                # Get active bets
+                my_active_back_bets = client.betting.my_bets(
+                    side=enums.BetSide.BACK, status=enums.BetRequestStatus.ACTIVE
+                )
+
+            else:
+                logger.exception(f"Issue creating bet request with the supplied params")
 
         else:
-            logger.exception(f"Issue creating bet request with the supplied params")
-
+            raise Exception(f"Could not find prices for fixture {fixture.fixture_id}")
     else:
-        raise Exception(f"Could not find prices for fixture {fixture.fixture_id}")
+        logger.info("No active markets could be found!")