+++ tests/unit/test_utils.py
@@ -1,6 +1,7 @@
 from betconnect import utils
 import pytest
 from uuid import UUID
+from betconnect import resources
 
 
 class TestUtils:
@@ -8,8 +9,14 @@ class TestUtils:
         assert utils.is_valid_uuid(uuid_to_test="123243") is False
         assert utils.is_valid_uuid("c9bf9e57-1685-4c89-bafb-ff5af830be8a")
 
-    def test_calculate_book_percentage(self):
-        raise NotImplementedError
+    def test_calculate_book_percentage(self, mock_selections_for_market_response):
+        all_selections = [
+            resources.SelectionsForMarket.create_from_dict(s)
+            for s in mock_selections_for_market_response[1]["data"]
+        ]
+        assert (
+            utils.calculate_book_percentage(market_selections=all_selections) == 1.222
+        )
 
     def test_parse_bet_request_id(self):
         assert isinstance(